// client/src/redux/actions/actionTypes.ts
export enum ACTION_TYPES {
    GET_EMPLOYEE_REQUEST = 'GET_EMPLOYEE_REQUEST',
    GET_EMPLOYEE_SUCCESS = 'GET_EMPLOYEE_SUCCESS',
    GET_EMPLOYEE_ERROR = 'GET_EMPLOYEE_ERROR',
    FETCH_EMPLOYEES_REQUEST = 'FETCH_EMPLOYEES_REQUEST',
    FETCH_EMPLOYEES_SUCCESS = 'FETCH_EMPLOYEES_SUCCESS',
    FETCH_EMPLOYEES_ERROR = 'FETCH_EMPLOYEES_ERROR',
    INSERT_EMPLOYEE_REQUEST = 'INSERT_EMPLOYEE_REQUEST',
    INSERT_EMPLOYEE_SUCCESS = 'INSERT_EMPLOYEE_SUCCESS',
    INSERT_EMPLOYEE_ERROR = 'INSERT_EMPLOYEE_ERROR',
    UPDATE_EMPLOYEE_REQUEST = 'UPDATE_EMPLOYEE_REQUEST',
    UPDATE_EMPLOYEE_SUCCESS = 'UPDATE_EMPLOYEE_SUCCESS',
    UPDATE_EMPLOYEE_ERROR = 'UPDATE_EMPLOYEE_ERROR',
    DELETE_EMPLOYEE_REQUEST = 'DELETE_EMPLOYEE_REQUEST',
    DELETE_EMPLOYEE_SUCCESS = 'DELETE_EMPLOYEE_SUCCESS',
    DELETE_EMPLOYEE_ERROR = 'DELETE_EMPLOYEE_ERROR',
};